Rename pages and routes
