+++ frontend/src/pages/Home/Home.tsx
@@ -1,0 +1,10 @@
+const Home = () => {
+  return (
+    <main>
+      <h1>Welcome to the Home Page</h1>
+      <p>This is the main landing page of the application.</p>
+    </main>
+  );
+};
+
+export default Home;
